add some table for client and owner in database.

+++ app/Http/Controllers/API/UserController.php
@@ -18,6 +18,7 @@ class UserController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', User::class);
         return UserResource::collection(User::all());
         // $user = User::all();
         // return response()->json(['users' => $user], 200);
@@ -31,6 +32,7 @@ class UserController extends Controller
     {
         //
         // The data is already validated by StoreUserRequest
+        $this->authorize('create', User::class);
         $validatedData = $request->validated();
 
         // Create the user
@@ -38,10 +40,10 @@ class UserController extends Controller
             'name'     => $validatedData['name'],
             'email'    => $validatedData['email'],
             'password' => Hash::make($validatedData['password']),
-            'phone_number' => $request->phone_number ?? null,
-            'national_id'  => $request->national_id ?? null,
-            'id_image'     => $request->id_image ?? null,
-            'location'     => $request->location ?? null,
+            'phone_number' => $validatedData->phone_number,
+            'national_id'  => $validatedData->national_id,
+            'id_image'     => $validatedData->id_image ?? null,
+            'location'     => $validatedData->location ?? null,
         ]);
 
         // You can return response as JSON or redirect
@@ -56,9 +58,14 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
-        //
-        $user= User::findorFail($id);
-        return response()->json(['user' => $user], 200);
+        $user = User::find($id);
+        if(!$user){
+            return response()->json(['message' => 'User of id ' . $id . ' not found'], 404);
+        }
+        $this->authorize('view', $user);
+        return new UserResource($user);
+
+        // return response()->json(['user' => $user], 200);
     }
 
     /**
@@ -67,12 +74,15 @@ class UserController extends Controller
     public function update(UpdateUserRequest $request, string $id)
     {
         //
-        $user = User::findorFail($id);
-
-        // The data is already validated by UpdateUserRequest
+        $user = User::find($id);
+        if(!$user){
+            return response()->json(['message' => 'User of id ' . $id . ' not found'], 404);
+        }
+        $this->authorize('update', $user);
         $validatedData = $request->validated();
-        $user->update($request->all());
+        $user->update($validatedData);
         return $user;
+
         // return response()->json(['message' => 'User updated'], 200);
     }
 
@@ -81,9 +91,13 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        //
-        $user = User::findorFail($id);
+        $user = User::find($id);
+        if(!$user){
+            return response()->json(['message' => 'User of id ' . $id . ' not found'], 404);
+        }
+        $this->authorize('delete', $user);
         $user->delete();
         return response()->json(['message' => 'User deleted'], 200);
+
     }
 }
